requirements.txt
black
flake8
matplotlib
numpy
Pillow
pre-commit
pyepics
p4p
